@@ -1,23 +1,23 @@
 name: bsv
 description: Bitcoin SV (BSV) library.
-version: 0.1.0
+version: 1.0.0
 homepage: "https://github.com/Dhuliang/flutter-bsv"
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  convert: ^2.1.1
-  pointycastle: ^2.0.0
-  bs58check: ^1.0.1
-  unorm_dart: ^0.1.2
+  convert: ^3.0.1
+  pointycastle: ^3.3.4
+  bs58check: ^1.0.2
+  unorm_dart: ^0.2.0
 
 dev_dependencies:
-  effective_dart: ^1.3.0
+  effective_dart: ^1.3.2
   flutter_test:
     sdk: flutter
 
