@@ -18,13 +18,13 @@ import 'package:bsv/extentsions/list.dart';
  */
 
 class Br {
-  Uint8List buf;
-  int pos;
+  Uint8List? buf;
+  int? pos;
 
   static const ERROR_NUMBER_TOO_LARGE =
       'number too large to retain precision - use readVarIntBn';
 
-  Br({Uint8List buf, int pos}) {
+  Br({Uint8List? buf, int? pos}) {
     this.buf = buf ?? Uint8List(0);
     this.pos = pos ?? 0;
   }
@@ -34,24 +34,24 @@ class Br {
   }
 
   bool eof() {
-    return this.pos >= this.buf.length;
+    return this.pos! >= this.buf!.length;
   }
 
-  Uint8List read([int len]) {
-    len = len ?? this.buf.length;
+  Uint8List read([int? len]) {
+    len = len ?? this.buf!.length;
     var start = this.pos;
-    var end = this.pos + len;
+    var end = this.pos! + len;
 
-    var newBuf = Uint8List.fromList(this.buf.slice(start, end));
+    var newBuf = Uint8List.fromList(this.buf!.slice(start, end));
     this.pos = end;
     return newBuf;
   }
 
-  List<int> readReverse([int len]) {
-    len = len ?? this.buf.length;
-    var buf = this.buf.sublist(this.pos, this.pos + len);
-    this.pos = this.pos + len;
-    var buf2 = List<int>(buf.length);
+  List<int> readReverse([int? len]) {
+    len = len ?? this.buf!.length;
+    var buf = this.buf!.sublist(this.pos!, this.pos! + len);
+    this.pos = this.pos! + len;
+    var buf2 = List<int>.filled(buf.length, 0, growable: false);
     for (var i = 0; i < buf2.length; i++) {
       buf2[i] = buf[buf.length - 1 - i];
     }
@@ -59,89 +59,89 @@ class Br {
   }
 
   int readUInt8() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Uint8List.fromList(this.buf).buffer);
-    var value = view.getUint8(this.pos);
-    this.pos = this.pos + 1;
+    var value = view.getUint8(this.pos!);
+    this.pos = this.pos! + 1;
     return value;
   }
 
   int readInt8() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Int8List.fromList(this.buf).buffer);
-    var value = view.getInt8(this.pos);
-    this.pos = this.pos + 1;
+    var value = view.getInt8(this.pos!);
+    this.pos = this.pos! + 1;
     return value;
   }
 
   int readUInt16BE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Uint16List.fromList(this.buf).buffer);
-    var value = view.getUint16(this.pos, Endian.big);
-    this.pos = this.pos + 2;
+    var value = view.getUint16(this.pos!, Endian.big);
+    this.pos = this.pos! + 2;
     return value;
   }
 
   int readInt16BE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Int8List.fromList(this.buf).buffer);
-    var value = view.getInt16(this.pos, Endian.big);
-    this.pos = this.pos + 2;
+    var value = view.getInt16(this.pos!, Endian.big);
+    this.pos = this.pos! + 2;
     return value;
   }
 
   int readUInt16LE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Uint16List.fromList(this.buf).buffer);
-    var value = view.getUint16(this.pos, Endian.little);
-    this.pos = this.pos + 2;
+    var value = view.getUint16(this.pos!, Endian.little);
+    this.pos = this.pos! + 2;
     return value;
   }
 
   int readInt16LE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Int8List.fromList(this.buf).buffer);
-    var value = view.getInt16(this.pos, Endian.little);
-    this.pos = this.pos + 2;
+    var value = view.getInt16(this.pos!, Endian.little);
+    this.pos = this.pos! + 2;
     return value;
   }
 
   int readUInt32BE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Uint32List.fromList(this.buf).buffer);
-    var value = view.getUint32(this.pos, Endian.big);
-    this.pos = this.pos + 4;
+    var value = view.getUint32(this.pos!, Endian.big);
+    this.pos = this.pos! + 4;
     return value;
   }
 
   int readInt32BE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Int32List.fromList(this.buf).buffer);
-    var value = view.getInt32(this.pos, Endian.big);
-    this.pos = this.pos + 4;
+    var value = view.getInt32(this.pos!, Endian.big);
+    this.pos = this.pos! + 4;
     return value;
   }
 
   int readUInt32LE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Uint32List.fromList(this.buf).buffer);
-    var value = view.getUint32(this.pos, Endian.little);
-    this.pos = this.pos + 4;
+    var value = view.getUint32(this.pos!, Endian.little);
+    this.pos = this.pos! + 4;
     return value;
   }
 
   int readInt32LE() {
-    var view = ByteData.view(this.buf.buffer);
+    var view = ByteData.view(this.buf!.buffer);
     // var view = ByteData.view(Int32List.fromList(this.buf).buffer);
-    var value = view.getInt32(this.pos, Endian.little);
-    this.pos = this.pos + 4;
+    var value = view.getInt32(this.pos!, Endian.little);
+    this.pos = this.pos! + 4;
     return value;
   }
 
   BigIntX readUInt64BEBn() {
-    var buf = this.buf.sublist(this.pos, this.pos + 8);
+    var buf = this.buf!.sublist(this.pos!, this.pos! + 8);
     var bn = BigIntX.fromBuffer(buf);
-    this.pos = this.pos + 8;
+    this.pos = this.pos! + 8;
     return bn;
   }
 
@@ -167,7 +167,7 @@ class Br {
         } else {
           throw ERROR_NUMBER_TOO_LARGE;
         }
-        break;
+      // break;
       default:
         return first;
     }
