@@ -23,20 +23,20 @@ class Tx {
   // This is defined on Interp, but Tx cannot depend on Interp - must redefine here
   static const int SCRIPT_ENABLE_SIGHASH_FORKID = 1 << 16;
 
-  int versionBytesNum;
-  VarInt txInsVi;
-  List<TxIn> txIns;
-  VarInt txOutsVi;
-  List<TxOut> txOuts;
-  int nLockTime;
+  int? versionBytesNum;
+  VarInt? txInsVi;
+  List<TxIn>? txIns;
+  VarInt? txOutsVi;
+  List<TxOut>? txOuts;
+  int? nLockTime;
 
   Tx({
-    int versionBytesNum,
-    VarInt txInsVi,
-    List<TxIn> txIns,
-    VarInt txOutsVi,
-    List<TxOut> txOuts,
-    int nLockTime,
+    int? versionBytesNum,
+    VarInt? txInsVi,
+    List<TxIn>? txIns,
+    VarInt? txOutsVi,
+    List<TxOut>? txOuts,
+    int? nLockTime,
   }) {
     this.versionBytesNum = versionBytesNum ?? 1;
     this.txInsVi = txInsVi ?? VarInt.fromNumber(0);
@@ -65,20 +65,20 @@ class Tx {
     );
   }
 
-  Map<String, Object> toJSON() {
+  Map<String, Object?> toJSON() {
     var txIns = [];
-    this.txIns.forEach((txIn) {
+    this.txIns!.forEach((txIn) {
       txIns.add(txIn.toJSON());
     });
     var txOuts = [];
-    this.txOuts.forEach((txOut) {
+    this.txOuts!.forEach((txOut) {
       txOuts.add(txOut.toJSON());
     });
     return {
       "versionBytesNum": this.versionBytesNum,
-      "txInsVi": this.txInsVi.toJSON(),
+      "txInsVi": this.txInsVi!.toJSON(),
       "txIns": txIns,
-      "txOutsVi": this.txOutsVi.toJSON(),
+      "txOutsVi": this.txOutsVi!.toJSON(),
       "txOuts": txOuts,
       "nLockTime": this.nLockTime
     };
@@ -87,61 +87,61 @@ class Tx {
   Tx fromBr(Br br) {
     this.versionBytesNum = br.readUInt32LE();
     this.txInsVi = VarInt(buf: br.readVarIntBuf());
-    var txInsNum = this.txInsVi.toNumber();
+    var txInsNum = this.txInsVi!.toNumber()!;
     this.txIns = [];
     for (var i = 0; i < txInsNum; i++) {
       var txIn = TxIn().fromBr(br);
-      this.txIns.add(txIn);
+      this.txIns!.add(txIn);
     }
     this.txOutsVi = VarInt(buf: br.readVarIntBuf());
-    var txOutsNum = this.txOutsVi.toNumber();
+    var txOutsNum = this.txOutsVi!.toNumber()!;
     this.txOuts = [];
     for (var i = 0; i < txOutsNum; i++) {
-      this.txOuts.add(TxOut().fromBr(br));
+      this.txOuts!.add(TxOut().fromBr(br));
     }
     this.nLockTime = br.readUInt32LE();
     return this;
   }
 
   factory Tx.fromBuffer(List<int> buf) {
-    return Tx().fromBr(Br(buf: buf));
+    return Tx().fromBr(Br(buf: buf as Uint8List?));
   }
 
   factory Tx.fromHex(String str) {
     return Tx().fromHex(str);
   }
 
-  Bw toBw([Bw bw]) {
+  Bw toBw([Bw? bw]) {
     if (bw == null) {
       bw = new Bw();
     }
-    bw.writeUInt32LE(this.versionBytesNum);
-    bw.write(this.txInsVi.buf.asUint8List());
-    for (var i = 0; i < this.txIns.length; i++) {
-      this.txIns[i].toBw(bw);
+    bw.writeUInt32LE(this.versionBytesNum!);
+    bw.write(this.txInsVi!.buf!.asUint8List());
+    for (var i = 0; i < this.txIns!.length; i++) {
+      this.txIns![i].toBw(bw);
     }
-    bw.write(this.txOutsVi.buf.asUint8List());
-    for (var i = 0; i < this.txOuts.length; i++) {
-      this.txOuts[i].toBw(bw);
+    bw.write(this.txOutsVi!.buf!.asUint8List());
+    for (var i = 0; i < this.txOuts!.length; i++) {
+      this.txOuts![i].toBw(bw);
     }
-    bw.writeUInt32LE(this.nLockTime);
+    bw.writeUInt32LE(this.nLockTime!);
     return bw;
   }
 
   // // https://github.com/Bitcoin-UAHF/spec/blob/master/replay-protected-sighash.md
   Hash hashPrevouts() {
     var bw = new Bw();
-    for (var txIn in this.txIns) {
+    for (var txIn in this.txIns!) {
       // var txIn = this.txIns[i];
-      bw.write(txIn.txHashBuf.asUint8List()); // outpoint (1/2)
-      bw.writeUInt32LE(txIn.txOutNum); // outpoint (2/2)
+      bw.write(txIn.txHashBuf!.asUint8List()); // outpoint (1/2)
+      bw.writeUInt32LE(txIn.txOutNum!); // outpoint (2/2)
     }
     return Hash.sha256Sha256(bw.toBuffer().asUint8List());
   }
 
   Hash hashSequence() {
     var bw = new Bw();
-    for (var txIn in this.txIns) {
+    for (var txIn in this.txIns!) {
       // var txIn = this.txIns[i]
       bw.writeUInt32LE(txIn.nSequence);
     }
@@ -150,7 +150,7 @@ class Tx {
 
   Hash hashOutputs() {
     var bw = new Bw();
-    for (var txOut in this.txOuts) {
+    for (var txOut in this.txOuts!) {
       // var txOut = this.txOuts[i]
       bw.write(txOut.toBuffer().asUint8List());
     }
@@ -164,12 +164,12 @@ class Tx {
    * normal because you're using OP_CODESEPARATORs, you know what to do.
    */
   List<int> sighash({
-    int nHashType,
-    int nIn,
-    Script subScript,
-    BigIntX valueBn,
-    int flags = 0,
-    HashCache hashCache,
+    required int nHashType,
+    int? nIn,
+    Script? subScript,
+    BigIntX? valueBn,
+    int? flags = 0,
+    HashCache? hashCache,
   }) {
     hashCache = hashCache ?? HashCache();
     var buf = this.sighashPreimage(
@@ -205,26 +205,26 @@ class Tx {
   // }
 
   List<int> sighashPreimage({
-    int nHashType,
-    int nIn,
-    Script subScript,
-    BigIntX valueBn,
-    int flags = 0,
-    HashCache hashCache,
+    required int nHashType,
+    int? nIn,
+    Script? subScript,
+    BigIntX? valueBn,
+    int? flags = 0,
+    HashCache? hashCache,
   }) {
     // start with UAHF part (Bitcoin SV)
     // https://github.com/Bitcoin-UAHF/spec/blob/master/replay-protected-sighash.md
     hashCache = hashCache ?? new HashCache();
     if ((nHashType & Sig.SIGHASH_FORKID != 0) &&
-        (flags & Tx.SCRIPT_ENABLE_SIGHASH_FORKID != 0)) {
-      var hashPrevouts = List.generate(32, (index) => 0);
-      var hashSequence = List.generate(32, (index) => 0);
-      var hashOutputs = List.generate(32, (index) => 0);
+        (flags! & Tx.SCRIPT_ENABLE_SIGHASH_FORKID != 0)) {
+      List<int>? hashPrevouts = List.generate(32, (index) => 0);
+      List<int>? hashSequence = List.generate(32, (index) => 0);
+      List<int>? hashOutputs = List.generate(32, (index) => 0);
 
       if (!(nHashType & Sig.SIGHASH_ANYONECANPAY != 0)) {
         hashPrevouts = hashCache.prevoutsHashBuf != null
             ? hashCache.prevoutsHashBuf
-            : hashCache.prevoutsHashBuf = this.hashPrevouts().data.toList();
+            : hashCache.prevoutsHashBuf = this.hashPrevouts().data!.toList();
       }
 
       if (!(nHashType & Sig.SIGHASH_ANYONECANPAY != 0) &&
@@ -232,32 +232,32 @@ class Tx {
           (nHashType & 0x1f) != Sig.SIGHASH_NONE) {
         hashSequence = hashCache.sequenceHashBuf != null
             ? hashCache.sequenceHashBuf
-            : hashCache.sequenceHashBuf = this.hashSequence().data.toList();
+            : hashCache.sequenceHashBuf = this.hashSequence().data!.toList();
       }
 
       if ((nHashType & 0x1f) != Sig.SIGHASH_SINGLE &&
           (nHashType & 0x1f) != Sig.SIGHASH_NONE) {
         hashOutputs = hashCache.outputsHashBuf != null
             ? hashCache.outputsHashBuf
-            : hashCache.outputsHashBuf = this.hashOutputs().data.toList();
+            : hashCache.outputsHashBuf = this.hashOutputs().data!.toList();
       } else if ((nHashType & 0x1f) == Sig.SIGHASH_SINGLE &&
-          nIn < this.txOuts.length) {
+          nIn! < this.txOuts!.length) {
         hashOutputs =
-            Hash.sha256Sha256(this.txOuts[nIn].toBuffer().asUint8List()).data;
+            Hash.sha256Sha256(this.txOuts![nIn].toBuffer().asUint8List()).data;
       }
 
       var bw = new Bw();
-      bw.writeUInt32LE(this.versionBytesNum);
-      bw.write(hashPrevouts.asUint8List());
-      bw.write(hashSequence.asUint8List());
-      bw.write(this.txIns[nIn].txHashBuf.asUint8List()); // outpoint (1/2);
-      bw.writeUInt32LE(this.txIns[nIn].txOutNum); // outpoint (2/2);
-      bw.writeVarIntNum(subScript.toBuffer().length);
+      bw.writeUInt32LE(this.versionBytesNum!);
+      bw.write(hashPrevouts!.asUint8List());
+      bw.write(hashSequence!.asUint8List());
+      bw.write(this.txIns![nIn!].txHashBuf!.asUint8List()); // outpoint (1/2);
+      bw.writeUInt32LE(this.txIns![nIn].txOutNum!); // outpoint (2/2);
+      bw.writeVarIntNum(subScript!.toBuffer().length);
       bw.write(subScript.toBuffer().asUint8List());
-      bw.writeUInt64LEBn(valueBn);
-      bw.writeUInt32LE(this.txIns[nIn].nSequence);
-      bw.write(hashOutputs.asUint8List());
-      bw.writeUInt32LE(this.nLockTime);
+      bw.writeUInt64LEBn(valueBn!);
+      bw.writeUInt32LE(this.txIns![nIn].nSequence);
+      bw.write(hashOutputs!.asUint8List());
+      bw.writeUInt32LE(this.nLockTime!);
       bw.writeUInt32LE(nHashType >> 0);
 
       return bw.toBuffer();
@@ -266,58 +266,59 @@ class Tx {
     // original bitcoin code follows - not related to UAHF (Bitcoin SV)
     var txcopy = this.cloneByBuffer();
 
-    subScript = new Script().fromBuffer(subScript.toBuffer().asUint8List());
+    subScript = new Script().fromBuffer(subScript!.toBuffer().asUint8List());
     subScript.removeCodeseparators();
 
-    for (var i = 0; i < txcopy.txIns.length; i++) {
-      txcopy.txIns[i] =
-          TxIn().fromBuffer(txcopy.txIns[i].toBuffer()).setScript(new Script());
+    for (var i = 0; i < txcopy.txIns!.length; i++) {
+      txcopy.txIns![i] = TxIn()
+          .fromBuffer(txcopy.txIns![i].toBuffer())
+          .setScript(new Script());
     }
 
-    txcopy.txIns[nIn] =
-        TxIn().fromBuffer(txcopy.txIns[nIn].toBuffer()).setScript(subScript);
+    txcopy.txIns![nIn!] =
+        TxIn().fromBuffer(txcopy.txIns![nIn].toBuffer()).setScript(subScript);
 
     if ((nHashType & 31) == Sig.SIGHASH_NONE) {
-      txcopy.txOuts.length = 0;
+      txcopy.txOuts!.length = 0;
       txcopy.txOutsVi = VarInt.fromNumber(0);
 
-      for (var i = 0; i < txcopy.txIns.length; i++) {
+      for (var i = 0; i < txcopy.txIns!.length; i++) {
         if (i != nIn) {
-          txcopy.txIns[i].nSequence = 0;
+          txcopy.txIns![i].nSequence = 0;
         }
       }
     } else if ((nHashType & 31) == Sig.SIGHASH_SINGLE) {
       // The SIGHASH_SINGLE bug.
       // https://bitcointalk.org/index.php?topic=260595.0
-      if (nIn > txcopy.txOuts.length - 1) {
+      if (nIn > txcopy.txOuts!.length - 1) {
         return hex.decode(
           '0000000000000000000000000000000000000000000000000000000000000001',
         );
       }
 
-      txcopy.txOuts.length = nIn + 1;
+      txcopy.txOuts!.length = nIn + 1;
       txcopy.txOutsVi = VarInt.fromNumber(nIn + 1);
 
-      for (var i = 0; i < txcopy.txOuts.length; i++) {
+      for (var i = 0; i < txcopy.txOuts!.length; i++) {
         if (i < nIn) {
-          txcopy.txOuts[i] = TxOut.fromProperties(
+          txcopy.txOuts![i] = TxOut.fromProperties(
             valueBn: new BigIntX.fromBuffer(hex.decode('ffffffffffffffff')),
             script: new Script(),
           );
         }
       }
 
-      for (var i = 0; i < txcopy.txIns.length; i++) {
+      for (var i = 0; i < txcopy.txIns!.length; i++) {
         if (i != nIn) {
-          txcopy.txIns[i].nSequence = 0;
+          txcopy.txIns![i].nSequence = 0;
         }
       }
     }
     // else, SIGHASH_ALL
 
     if (nHashType & Sig.SIGHASH_ANYONECANPAY != 0) {
-      txcopy.txIns[0] = txcopy.txIns[nIn];
-      txcopy.txIns.length = 1;
+      txcopy.txIns![0] = txcopy.txIns![nIn];
+      txcopy.txIns!.length = 1;
       txcopy.txInsVi = VarInt.fromNumber(1);
     }
 
@@ -363,13 +364,13 @@ class Tx {
 
   // This  returns a signature but does not update any inputs
   Sig sign({
-    KeyPair keyPair,
+    KeyPair? keyPair,
     int nHashType = Sig.SIGHASH_ALL | Sig.SIGHASH_FORKID,
-    int nIn,
-    Script subScript,
-    BigIntX valueBn,
+    int? nIn,
+    Script? subScript,
+    BigIntX? valueBn,
     int flags = Tx.SCRIPT_ENABLE_SIGHASH_FORKID,
-    HashCache hashCache,
+    HashCache? hashCache,
   }) {
     var hashBuf = this.sighash(
       nHashType: nHashType,
@@ -390,7 +391,7 @@ class Tx {
       hashBuf: hashBuf,
       keyPair: keyPair,
       endian: Endian.little,
-    );
+    )!;
     sig.nHashType = nHashType;
     return sig;
   }
@@ -409,18 +410,18 @@ class Tx {
   // }
 
   // This  takes a signature as input and does not parse any inputs
-  bool verify({
-    Sig sig,
-    PubKey pubKey,
-    int nIn,
-    Script subScript,
+  bool? verify({
+    required Sig sig,
+    PubKey? pubKey,
+    int? nIn,
+    Script? subScript,
     bool enforceLowS = false,
-    BigIntX valueBn,
-    int flags = Tx.SCRIPT_ENABLE_SIGHASH_FORKID,
-    HashCache hashCache,
+    BigIntX? valueBn,
+    int? flags = Tx.SCRIPT_ENABLE_SIGHASH_FORKID,
+    HashCache? hashCache,
   }) {
     var hashBuf = this.sighash(
-      nHashType: sig.nHashType,
+      nHashType: sig.nHashType!,
       nIn: nIn,
       subScript: subScript,
       valueBn: valueBn,
@@ -480,9 +481,9 @@ class Tx {
 
   Tx addTxIn({
     dynamic data,
-    int txOutNum,
-    Script script,
-    int nSequence,
+    int? txOutNum,
+    Script? script,
+    int? nSequence,
   }) {
     var txIn;
     if (data is TxIn) {
@@ -495,20 +496,20 @@ class Tx {
         script: script,
       ).setScript(script);
     }
-    this.txIns.add(txIn);
-    this.txInsVi = VarInt.fromNumber(this.txInsVi.toNumber() + 1);
+    this.txIns!.add(txIn);
+    this.txInsVi = VarInt.fromNumber(this.txInsVi!.toNumber()! + 1);
     return this;
   }
 
-  Tx addTxOut({dynamic data, Script script}) {
+  Tx addTxOut({dynamic data, Script? script}) {
     var txOut;
     if (data is TxOut) {
       txOut = data;
     } else {
-      txOut = new TxOut(valueBn: data).setScript(script);
+      txOut = new TxOut(valueBn: data).setScript(script!);
     }
-    this.txOuts.add(txOut);
-    this.txOutsVi = VarInt.fromNumber(this.txOutsVi.toNumber() + 1);
+    this.txOuts!.add(txOut);
+    this.txOutsVi = VarInt.fromNumber(this.txOutsVi!.toNumber()! + 1);
     return this;
   }
 
@@ -517,7 +518,7 @@ class Tx {
    * Analagous to bitcoind's IsCoinBase  in transaction.h
    */
   bool isCoinbase() {
-    return this.txIns.length == 1 && this.txIns[0].hasNullInput();
+    return this.txIns!.length == 1 && this.txIns![0].hasNullInput();
   }
 
   // ignore: slash_for_doc_comments
@@ -525,27 +526,27 @@ class Tx {
    * BIP 69 sorting. Be sure to sign after sorting.
    */
   Tx sort() {
-    this.txIns.sort((first, second) {
+    this.txIns!.sort((first, second) {
       // return new Br(buf: first.txHashBuf)
       //         .readReverse()
       //         .compare(new Br(buf: second.txHashBuf).readReverse()) ??
       //     first.txOutNum - second.txOutNum;
-      var a = new Br(buf: first.txHashBuf.asUint8List()).readReverse();
-      var b = new Br(buf: second.txHashBuf.asUint8List()).readReverse();
+      var a = new Br(buf: first.txHashBuf!.asUint8List()).readReverse();
+      var b = new Br(buf: second.txHashBuf!.asUint8List()).readReverse();
 
       var result = a.compareTo(b);
       // 如果是0, 则返回first.txOutNum - second.txOutNum;
       // 否则返回相对应的排序
-      if (result == 0) return first.txOutNum - second.txOutNum;
+      if (result == 0) return first.txOutNum! - second.txOutNum!;
       return result;
     });
 
-    this.txOuts.sort((first, second) {
-      var result = first.valueBn.sub(second.valueBn).toNumber();
+    this.txOuts!.sort((first, second) {
+      var result = first.valueBn!.sub(second.valueBn!).toNumber();
       if (result != 0) return result;
 
-      var a = first.script.toBuffer();
-      var b = second.script.toBuffer();
+      var a = first.script!.toBuffer();
+      var b = second.script!.toBuffer();
       return a.compareTo(b);
     });
 
