@@ -5,6 +5,7 @@ void main() {
   group('OpCode', () {
     test('should create a new OpCode', () {
       var opCode = new OpCode(number: 5);
+      // ignore: unnecessary_null_comparison
       expect(opCode.number != null, true);
     });
 
@@ -25,34 +26,34 @@ void main() {
     });
 
     test('should convert to a number with this handy syntax', () {
-      expect(new OpCode().fromString('OP_0').toNumber(), 0);
+      expect(new OpCode.fromString('OP_0').toNumber(), 0);
 
-      expect(new OpCode().fromString('OP_16').toNumber(), 96);
+      expect(new OpCode.fromString('OP_16').toNumber(), 96);
 
-      expect(new OpCode().fromString('OP_NOP').toNumber(), 97);
+      expect(new OpCode.fromString('OP_NOP').toNumber(), 97);
     });
 
     group('#fromNumber', () {
       test('should work for 0', () {
-        expect(OpCode().fromNumber(0).number, 0);
+        expect(OpCode.fromNumber(0).number, 0);
       });
     });
 
     group('#toNumber', () {
       test('should work for 0', () {
-        expect(OpCode().fromNumber(0).toNumber(), 0);
+        expect(OpCode.fromNumber(0).toNumber(), 0);
       });
     });
 
     group('#fromString', () {
       test('should work for OP_0', () {
-        expect(OpCode().fromString('OP_0').number, 0);
+        expect(OpCode.fromString('OP_0').number, 0);
       });
     });
 
     group('#toString', () {
       test('should work for OP_0', () {
-        expect(OpCode().fromString('OP_0').toString(), 'OP_0');
+        expect(OpCode.fromString('OP_0').toString(), 'OP_0');
       });
     });
 
