@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:bsv/random.dart';
-import 'package:bsv/common_util.dart';
+import 'package:bsv/extentsions/list.dart';
 
 void main() {
   group("RandomBytes", () {
@@ -18,8 +18,8 @@ void main() {
         var bytes1 = RandomBytes.getRandomBuffer(32);
         var bytes2 = RandomBytes.getRandomBuffer(32);
 
-        var hex1 = CommonUtil.list2hex(bytes1);
-        var hex2 = CommonUtil.list2hex(bytes2);
+        var hex1 = bytes1.toHex();
+        var hex2 = bytes2.toHex();
 
         expect(hex1 == hex2, false);
         expect(hex1.length, 64);
@@ -31,7 +31,7 @@ void main() {
 
         for (var i = 0; i < 100; i++) {
           var bytes = RandomBytes.getRandomBuffer(8);
-          hexs.add(CommonUtil.list2hex(bytes));
+          hexs.add(bytes.toHex());
         }
 
         for (var i = 0; i < 100; i++) {
