@@ -92,9 +92,9 @@ void main() {
       test('should set all outputs from a tx', () {
         var txOutMap = new TxOutMap().setTx(tx);
         var txHashBuf = tx.hash();
-        var txOut = tx.txOuts[0];
+        var txOut = tx.txOuts![0];
 
-        expect(txOutMap.get(txHashBuf.data, 0).toHex(), txOut.toHex());
+        expect(txOutMap.get(txHashBuf.data!, 0).toHex(), txOut.toHex());
       });
     });
   });
